# openapi/orbithr-v1.yaml
openapi: 3.1.0
info:
  title: OrbitHR API
  version: 1.0.0
  description: Authenticated, tenant-isolated API for OrbitHR web and Android clients.
servers:
  - url: http://localhost:3001/api/v1
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      operationId: getEmployees
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 100}}
        - {in: query, name: search, schema: {type: string}}
      responses:
        '200': {description: Permission and tenant-scoped employee page}
...
  /search:
    get:
      operationId: universalAuthorizedSearch
      parameters:
        - {in: query, name: q, required: true, schema: {type: string, minLength: 2, maxLength: 80}}
      responses:
        '200': {description: Permission and employee-scope filtered search results}
...
  /security/sessions/{id}:
    delete:
      operationId: revokeSession
      parameters:
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Session revoked}
...
  /employee-documents/{id}/review:
    post:
      operationId: reviewEmployeeDocument
      parameters:
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Document verification status updated}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...